landspace_web/requirements.txt
click==6.7
coreapi==2.1.1
coreapi-cli==1.0.6
Django==1.10.5
django-webpack-loader==0.4.1
djangorestframework==3.5.3
httplib2==0.9.2
itypes==1.1.0
oauth2==1.9.0.post1
psycopg2==2.6.2
python-openid==2.2.5
requests==2.12.4
uritemplate==3.0.0
